+++ @apicize/toolkit/src/controls/viewers/results-viewer.tsx
@@ -36,8 +36,6 @@ export const ResultsViewer = observer((props: {
         return null
     }
 
-    const isRunning = requestExecution.running
-
     const requestOrGroupId = workspace.active.id
     const selectedExecution = requestExecution.resultMenu[requestExecution.resultIndex]
     const executionResultId = selectedExecution?.executionResultId
@@ -60,7 +58,7 @@ export const ResultsViewer = observer((props: {
         ? executionResult as WorkbookExecutionGroup
         : undefined
 
-    const disableOtherPanels = executionRequestResult?.response === undefined
+    const disableOtherPanels = (!executionRequestResult?.response)
 
     const longTextInResponse = (executionRequestResult?.response?.body?.text?.length ?? 0)
         > MAX_TEXT_RENDER_LENGTH
@@ -97,7 +95,6 @@ export const ResultsViewer = observer((props: {
             <ToggleButtonGroup
                 orientation='vertical'
                 exclusive
-                disabled={isRunning}
                 onChange={handlePanelChanged}
                 value={panel}
                 sx={{ marginRight: '24px' }}
@@ -110,7 +107,6 @@ export const ResultsViewer = observer((props: {
             </ToggleButtonGroup>
             <Box sx={{ overflow: 'hidden', flexGrow: 1, bottom: '0', position: 'relative' }}>
                 <Box position='relative' width='100%' height='100%'>
-                    <Box position='absolute' display={isRunning ? 'block' : 'none'} top={0} left={0} width='100%' height='100%' zIndex={9999} sx={{backgroundColor: '#00000080'}} />
                     {
                         panel === 'Info' ? <ResultInfoViewer requestOrGroupId={requestOrGroupId} executionResultId={executionResultId} />
                             : panel === 'Headers' ? <ResponseHeadersViewer requestOrGroupId={requestOrGroupId} executionResultId={executionResultId} />