+++ app/src/providers/apicize-settings.provider.tsx
@@ -1,12 +1,5 @@
-import { StoredGlobalSettings } from "@apicize/lib-typescript";
-import { SupportedColorScheme } from "@mui/material";
-import { makeAutoObservable } from "mobx";
-import { createContext, ReactNode, useContext } from "react";
-import * as core from '@tauri-apps/api/core'
-import * as path from '@tauri-apps/api/path'
+import { ReactNode } from "react";
 import { ApicizeSettingsContext, ApicizeSettingsStore } from "@apicize/toolkit";
-
-
 
 export function ApicizeSettingsProvider({
     store, children